--- frontend/src/assets/pages/QuestionPage.tsx
@@ -1,9 +1,0 @@
-import QuestionTable from "../components/question/QuestionTable";
-
-const QuestionPage = () => {
-  return (
-    <QuestionTable />
-  );
-}
-
-export default QuestionPage;
