+++ frontend/src/question/DescriptionPopup.tsx
